+++ files/private-chef-cookbooks/private-chef/libraries/private_chef.rb
@@ -321,7 +321,6 @@ module PrivateChef
       PrivateChef["couchdb"]["bind_address"] ||= PrivateChef["default_listen_address"]
       PrivateChef["rabbitmq"]["node_ip_address"] ||= PrivateChef["default_listen_address"]
       PrivateChef["redis_lb"]["listen"] ||= PrivateChef["default_listen_address"]
-      PrivateChef["nginx"]["enable_ipv6"] ||= PrivateChef["use_ipv6"]
       PrivateChef["opscode_solr"]["ip_address"] ||= PrivateChef["default_listen_address"]
       PrivateChef["opscode_webui"]["worker_processes"] ||= 2
       PrivateChef["postgresql"]["listen_address"] ||= '*' #PrivateChef["default_listen_address"]
@@ -339,7 +338,6 @@ module PrivateChef
 
     def gen_frontend
       PrivateChef[:role] = "frontend"
-      PrivateChef["nginx"]["enable_ipv6"] ||= PrivateChef["use_ipv6"]
       PrivateChef["bookshelf"]["enable"] ||= false
       PrivateChef["bookshelf"]["vip"] ||= PrivateChef["backend_vips"]["ipaddress"]
       PrivateChef["couchdb"]["enable"] ||= false
@@ -418,6 +416,8 @@ module PrivateChef
         PrivateChef["use_ipv6"] = true
         PrivateChef["default_listen_address"] = "::"
       end
+
+      PrivateChef["nginx"]["enable_ipv6"] ||= PrivateChef["use_ipv6"]
 
       case PrivateChef['topology']
       when "standalone","manual"