+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 * Introduce pluggable HA architecture as an alternative to DRBD
 * [OC-11673] Tune PostgreSQL keepalive timeouts
 * [OC-10117] opscode-solr4 accepts Java-like memory attributes
+* [OC-11668] enable ipv6 in standalone mode
 
 ### openssl 1.0.1h
 * Allow ['lb']['upstream'] to have a custom setting