+++ CHANGELOG.md
@@ -81,6 +81,9 @@
 * Use SSL port for lb_internal if non-SSL is disabled
 * Restart redis_lb immediately during reconfigure
 
+### oc-id 0.3.3
+* Add Chef Identity Service (oc-id)
+
 ## 11.1.6 (2014-06-05)
 
 ### openssl 1.0.1h